refactor : ErrorMessage로 이름 변경
